requirements.txt
# Core dependencies
paddleocr>=2.7.0
paddlepaddle>=2.5.0
opencv-python>=4.8.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
numpy>=1.24.0

# Language detection and translation
langdetect>=1.0.9
googletrans==4.0.0rc1

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
